#183 move tic tac toe state to submodule
